fix(lint): Auto-fix 541 Python lint issues with ruff
- Sorted __all__ exports (RUF022)
- Fixed import sorting (I001)
- Removed unused imports (F401)
- Updated type annotations (UP035, UP006, UP045)
  - List/Dict/Tuple → list/dict/tuple
  - Optional[X] → X | None
- Fixed unused variables (F841)
- Applied various style fixes (SIM102, B007, RUF005, etc.)

50 remaining issues require manual fixes or noqa comments.

+++ agenkit/evaluation/bayesian_optimizer.py
@@ -123,7 +123,7 @@ class BayesianOptimizer(Optimizer):
         for i, name in enumerate(self.param_names):
             spec = self.search_space.parameters[name]
             if spec["type"] == "integer":
-                config[name] = int(round(vector[i]))
+                config[name] = round(vector[i])
             else:
                 config[name] = float(vector[i])
